gedie_chat/robot/requirements.txt
robotframework==7.0
telethon==1.34.0
python-dotenv==1.0.1
Pillow==10.3.0
